# zstd: Copy literal in 16 byte blocks when possible