@WojciechMula 2022-05-12
A funny fact is that I wanted to pick that problem and planned to ask you where to start. :)